requirements.txt
argh==0.26.2
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
future==0.17.1
idna==2.8
pathtools==0.1.2
pycparser==2.19
python-telegram-bot==11.1.0
PyYAML==3.13
requests==2.21.0
six==1.12.0
urllib3==1.24.1
watchdog==0.9.0
